Rename: main 디렉토리로 main 관련 js파일 전부 옮김
