+++ js/main/dayFilter/addClickEvent.js
