+++ js/main/dayFilter/clearDayFilter.js
