+++ js/main/dayFilter/replaceDayContents.js
