+++ js/main/dayFilter/setDayFilter.js
