+++ js/main/preview/clearSlideBtn.js
