+++ js/main/preview/drawSlideBtn.js
