+++ js/main/webtoonComponent/clearWebtoonContents.js
